+++ lib/model_viewer_server.dart
@@ -1,0 +1,18 @@
+library model_viewer_server;
+
+import "dart:io";
+import 'dart:json';
+import 'dart:uri';
+import "dart:utf";
+import "dart:math";
+import "package:logging/logging.dart";
+import 'package:objectory/objectory_console.dart';
+
+part 'src/server/model_object.dart';
+part 'src/server/model_service.dart';
+part 'src/server/server.dart';
+part 'src/server/server_configuration.dart';
+part 'src/server/static_file_server.dart';
+part 'src/server/url_shortener.dart';
+part 'src/server/websocket_commands.dart';
+part 'src/server/websocket_server.dart';
