+++ lib/src/server/model_object.dart
@@ -1,0 +1,12 @@
+part of model_viewer_server;
+
+class ModelObject extends PersistentObject {
+  String get modelState => getProperty('modelState');
+  set modelState(String value) => setProperty('modelState', value);
+}
+
+void registerClasses() {
+  objectory.registerClass('ModelObject', () => new ModelObject());
+}
+
+ObjectoryQueryBuilder get $ModelObject => new ObjectoryQueryBuilder('ModelObject');
