+++ lib/src/server/websocket_commands.dart
@@ -1,0 +1,7 @@
+part of model_viewer_server;
+
+class WebSocketCommands {
+  static final COMMAND = "command";
+  static final STORE = "store";
+  static final LOAD = "load";
+}
